Add timeout logic for LED color changes when distance less than 15 cm

+++ WEMOS_RECIEVER/src/wifi_reciever.cpp
@@ -57,7 +57,7 @@ void setup() {
 }
 
 void loop() {
-  unsigned long currentTime = millis();
+  uint32_t currentTime = millis();
   if(currentTime - previousTime >= timeout) {
     if(WiFi.status() == WL_CONNECTED) {
       Serial.println("Starting distance and LED data requests...");
